fix: improve equatable comparison of QueryConstraint
