+++ Sources/ParseSwift/Extensions/Data.swift
@@ -1,5 +1,5 @@
 //
-//  Data+hexString.swift
+//  Data.swift
 //  ParseSwift
 //
 //  Created by Corey Baker on 2/14/21.
