+++ Sources/ParseSwift/InternalObjects/BaseParseInstallation.swift
